+++ compile/competent.js
@@ -7,8 +7,8 @@ const util = require('util');
 const path = require('path');
 const fs = require('fs');             
 const aa = /acit|ex(?:s|g|n|p|$)|rph|ows|mnc|ntw|ine[ch]|zoo|^ord/i;
-let G = a => `${a}`.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;").replace(/"/g, "&quot;"), ba = a => 40 < `${a}`.length || -1 != `${a}`.indexOf("\n") || -1 !== `${a}`.indexOf("<");
-const H = {};
+let F = a => `${a}`.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;").replace(/"/g, "&quot;"), ba = a => 40 < `${a}`.length || -1 != `${a}`.indexOf("\n") || -1 !== `${a}`.indexOf("<");
+const G = {};
 function ca(a) {
   const b = {...a.attributes, children:a.children};
   a = a.nodeName.defaultProps;
@@ -51,8 +51,8 @@ function ca(a) {
         {
           let t = "";
           for (var u in m) {
-            let y = m[u];
-            null != y && (t && (t += " "), t += H[u] || (H[u] = u.replace(/([A-Z])/g, "-$1").toLowerCase()), t += ": ", t += y, "number" == typeof y && !1 === aa.test(u) && (t += "px"), t += ";");
+            let z = m[u];
+            null != z && (t && (t += " "), t += G[u] || (G[u] = u.replace(/([A-Z])/g, "-$1").toLowerCase()), t += ": ", t += z, "number" == typeof z && !1 === aa.test(u) && (t += "px"), t += ";");
           }
           m = t || void 0;
         }
@@ -74,13 +74,13 @@ function ca(a) {
             }
             "option" == b && g == m && (u = "selected ");
           }
-          return `${u}${h}="${G(m)}"`;
+          return `${u}${h}="${F(m)}"`;
         }
       }
     }
   }).filter(Boolean), v:k, l:g};
 };
-const J = [], ea = /^(area|base|br|col|embed|hr|img|input|link|meta|param|source|track|wbr)$/, fa = /^(a|abbr|acronym|audio|b|bdi|bdo|big|br|button|canvas|cite|code|data|datalist|del|dfn|em|embed|i|iframe|img|input|ins|kbd|label|map|mark|meter|noscript|object|output|picture|progress|q|ruby|s|samp|slot|small|span|strong|sub|sup|svg|template|textarea|time|u|tt|var|video|wbr)$/, N = (a, b = {}) => {
+const K = [], ea = /^(area|base|br|col|embed|hr|img|input|link|meta|param|source|track|wbr)$/, fa = /^(a|abbr|acronym|audio|b|bdi|bdo|big|br|button|canvas|cite|code|data|datalist|del|dfn|em|embed|i|iframe|img|input|ins|kbd|label|map|mark|meter|noscript|object|output|picture|progress|q|ruby|s|samp|slot|small|span|strong|sub|sup|svg|template|textarea|time|u|tt|var|video|wbr)$/, N = (a, b = {}) => {
   const {addDoctype:c, pretty:d} = b;
   a = M(a, b, {});
   return c ? `<!doctype html>${d ? "\n" : ""}${a}` : a;
@@ -89,32 +89,32 @@ function M(a, b = {}, c = {}, d = !1, e = !1, f) {
   if (null == a || "boolean" == typeof a) {
     return "";
   }
-  const {pretty:g = !1, shallow:k = !1, renderRootComponent:p = !1, shallowHighOrder:h = !1, sortAttributes:m, allAttributes:u, xml:t, initialPadding:y = 0, closeVoidTags:Q = !1} = b;
-  let {lineLength:B = 40} = b;
-  B -= y;
-  let {nodeName:l, attributes:K = {}} = a;
-  var q = ["textarea", "pre"].includes(l), v = " ".repeat(y);
-  const D = "string" == typeof g ? g : `  ${v}`;
+  const {pretty:g = !1, shallow:k = !1, renderRootComponent:p = !1, shallowHighOrder:h = !1, sortAttributes:m, allAttributes:u, xml:t, initialPadding:z = 0, closeVoidTags:Q = !1} = b;
+  let {lineLength:C = 40} = b;
+  C -= z;
+  let {nodeName:l, attributes:L = {}} = a;
+  var q = ["textarea", "pre"].includes(l), w = " ".repeat(z);
+  const x = "string" == typeof g ? g : `  ${w}`;
   if ("object" != typeof a && !l) {
-    return G(a);
+    return F(a);
   }
   if ("function" == typeof l) {
     if (!k || !d && p) {
-      return v = ca(a), l.prototype && "function" == typeof l.prototype.render ? (q = new l(v, c), q._disable = q.__x = !0, q.props = v, q.context = c, l.getDerivedStateFromProps ? q.state = {...q.state, ...l.getDerivedStateFromProps(q.props, q.state)} : q.componentWillMount && q.componentWillMount(), v = q.render(q.props, q.state, q.context), q.getChildContext && (c = {...c, ...q.getChildContext()})) : v = l(v, c), M(v, b, c, h, e, f);
+      return w = ca(a), l.prototype && "function" == typeof l.prototype.render ? (q = new l(w, c), q._disable = q.__x = !0, q.props = w, q.context = c, l.getDerivedStateFromProps ? q.state = {...q.state, ...l.getDerivedStateFromProps(q.props, q.state)} : q.componentWillMount && q.componentWillMount(), w = q.render(q.props, q.state, q.context), q.getChildContext && (c = {...c, ...q.getChildContext()})) : w = l(w, c), M(w, b, c, h, e, f);
     }
     l = l.displayName || l !== Function && l.name || ha(l);
   }
   let r = "";
-  ({A:E, v:d, l:f} = da(K, l, {allAttributes:u, xml:t, w:e, sort:m, l:f}));
+  ({A:E, v:d, l:f} = da(L, l, {allAttributes:u, xml:t, w:e, sort:m, l:f}));
   if (g) {
-    let z = `<${l}`.length;
-    r = E.reduce((x, F) => {
-      const L = z + 1 + F.length;
-      if (L > B) {
-        return z = D.length, `${x}\n${D}${F}`;
+    let y = `<${l}`.length;
+    r = E.reduce((v, I) => {
+      const J = y + 1 + I.length;
+      if (J > C) {
+        return y = x.length, `${v}\n${x}${I}`;
       }
-      z = L;
-      return `${x} ${F}`;
+      y = J;
+      return `${v} ${I}`;
     }, "");
   } else {
     r = E.length ? " " + E.join(" ") : "";
@@ -125,45 +125,45 @@ function M(a, b = {}, c = {}, d = !1, e = !1, f) {
   }
   var E = `${l}`.match(ea);
   Q && E && (r = r.replace(/>$/, " />"));
-  let w = [];
+  let A = [];
   if (d) {
-    !q && g && (ba(d) || d.length + O(r) > B) && (d = "\n" + D + `${d}`.replace(/(\n+)/g, "$1" + (D || "\t"))), r += d;
+    !q && g && (ba(d) || d.length + O(r) > C) && (d = "\n" + x + `${d}`.replace(/(\n+)/g, "$1" + (x || "\t"))), r += d;
   } else {
     if (a.children) {
-      let z = g && r.includes("\n");
-      w = a.children.map(x => {
-        if (null != x && !1 !== x && (x = M(x, b, c, !0, "svg" == l ? !0 : "foreignObject" == l ? !1 : e, f))) {
-          return g && x.length + O(r) > B && (z = !0), x;
+      let y = g && r.includes("\n");
+      A = a.children.map(v => {
+        if (null != v && !1 !== v && (v = M(v, b, c, !0, "svg" == l ? !0 : "foreignObject" == l ? !1 : e, f))) {
+          return g && v.length + O(r) > C && (y = !0), v;
         }
       }).filter(Boolean);
-      if (g && z && !q) {
-        for (a = w.length; a--;) {
-          w[a] = "\n" + D + `${w[a]}`.replace(/(\n+)/g, "$1" + (D || "\t"));
+      if (g && y && !q) {
+        for (a = A.length; a--;) {
+          A[a] = "\n" + x + `${A[a]}`.replace(/(\n+)/g, "$1" + (x || "\t"));
         }
       }
     }
   }
-  if (w.length) {
-    r += w.join("");
+  if (A.length) {
+    r += A.join("");
   } else {
     if (t) {
       return r.substring(0, r.length - 1) + " />";
     }
   }
-  E || (a = w[w.length - 1], `${l}`.match(fa) && (a ? !/>$/.test(a) : 1) || q || !g || !r.includes("\n") || (r += `\n${v}`), r += `</${l}>`);
+  E || (a = A[A.length - 1], `${l}`.match(fa) && (a ? !/>$/.test(a) : 1) || q || !g || !r.includes("\n") || (r += `\n${w}`), r += `</${l}>`);
   return r;
 }
 function ha(a) {
   var b = (Function.prototype.toString.call(a).match(/^\s*function\s+([^( ]+)/) || "")[1];
   if (!b) {
     b = -1;
-    for (let c = J.length; c--;) {
-      if (J[c] === a) {
+    for (let c = K.length; c--;) {
+      if (K[c] === a) {
         b = c;
         break;
       }
     }
-    0 > b && (b = J.push(a) - 1);
+    0 > b && (b = K.push(a) - 1);
     b = `UnnamedComponent${b}`;
   }
   return b;
@@ -202,9 +202,9 @@ const ma = (a, b) => {
   b[c] = "true" == d ? !0 : "false" == d ? !1 : /^\d+$/.test(d) ? parseInt(d, 10) : d;
   return b;
 }, {});
-var ta = stream;
-const {Transform:ua, Writable:va} = stream;
-function wa(a) {
+var ua = stream;
+const {Transform:va, Writable:wa} = stream;
+function xa(a) {
   if ("object" != typeof a) {
     return !1;
   }
@@ -213,7 +213,7 @@ function wa(a) {
   const d = -1 != ["string", "function"].indexOf(typeof c);
   return a && d;
 }
-const xa = (a, b) => {
+const ya = (a, b) => {
   if (!(b instanceof Error)) {
     throw b;
   }
@@ -227,32 +227,32 @@ const xa = (a, b) => {
   b.stack = a.substr(0, c);
   throw b;
 };
-const {homedir:ya} = os;
-const za = /\s+at.*(?:\(|\s)(.*)\)?/, Aa = /^(?:(?:(?:node|(?:internal\/[\w/]*|.*node_modules\/(?:IGNORED_MODULES)\/.*)?\w+)\.js:\d+:\d+)|native)/, Ba = ya(), P = a => {
-  const {pretty:b = !1, ignoredModules:c = ["pirates"]} = {}, d = c.join("|"), e = new RegExp(Aa.source.replace("IGNORED_MODULES", d));
+const {homedir:za} = os;
+const Aa = /\s+at.*(?:\(|\s)(.*)\)?/, Ba = /^(?:(?:(?:node|(?:internal\/[\w/]*|.*node_modules\/(?:IGNORED_MODULES)\/.*)?\w+)\.js:\d+:\d+)|native)/, Ca = za(), P = a => {
+  const {pretty:b = !1, ignoredModules:c = ["pirates"]} = {}, d = c.join("|"), e = new RegExp(Ba.source.replace("IGNORED_MODULES", d));
   return a.replace(/\\/g, "/").split("\n").filter(f => {
-    f = f.match(za);
+    f = f.match(Aa);
     if (null === f || !f[1]) {
       return !0;
     }
     f = f[1];
     return f.includes(".app/Contents/Resources/electron.asar") || f.includes(".app/Contents/Resources/default_app.asar") ? !1 : !e.test(f);
-  }).filter(f => f.trim()).map(f => b ? f.replace(za, (g, k) => g.replace(k, k.replace(Ba, "~"))) : f).join("\n");
+  }).filter(f => f.trim()).map(f => b ? f.replace(Aa, (g, k) => g.replace(k, k.replace(Ca, "~"))) : f).join("\n");
 };
-const Ca = (a, b = 0, c = !1) => {
+const Da = (a, b = 0, c = !1) => {
   if (0 === b && !c) {
     return a;
   }
   a = a.split("\n", c ? b + 1 : void 0);
   return c ? a[a.length - 1] : a.slice(b).join("\n");
-}, Da = (a, b = !1) => Ca(a, 2 + (b ? 1 : 0)), Ea = a => {
+}, Ea = (a, b = !1) => Da(a, 2 + (b ? 1 : 0)), Fa = a => {
   ({callee:{caller:a}} = a);
   return a;
 };
-function Fa(a, b, c = !1) {
+function Ga(a, b, c = !1) {
   return function(d) {
-    var e = Ea(arguments), {stack:f} = Error();
-    const g = Ca(f, 2, !0), k = (f = d instanceof Error) ? d.message : d;
+    var e = Fa(arguments), {stack:f} = Error();
+    const g = Da(f, 2, !0), k = (f = d instanceof Error) ? d.message : d;
     e = [`Error: ${k}`, ...null !== e && a === e || c ? [b] : [g, b]].join("\n");
     e = P(e);
     return Object.assign(f ? d : Error(), {message:k, stack:e});
@@ -260,17 +260,17 @@ function Fa(a, b, c = !1) {
 }
 ;function R(a) {
   var {stack:b} = Error();
-  const c = Ea(arguments);
-  b = Da(b, a);
-  return Fa(c, b, a);
+  const c = Fa(arguments);
+  b = Ea(b, a);
+  return Ga(c, b, a);
 }
-;const Ga = (a, b) => {
+;const Ha = (a, b) => {
   b.once("error", c => {
     a.emit("error", c);
   });
   return b;
 };
-class Ha extends va {
+class Ia extends wa {
   constructor(a) {
     const {binary:b = !1, rs:c = null, ...d} = a || {}, {u:e = R(!0), proxyError:f} = a || {}, g = (k, p) => e(p);
     super(d);
@@ -292,7 +292,7 @@ class Ha extends va {
         }
         p(h);
       });
-      c && Ga(this, c).pipe(this);
+      c && Ha(this, c).pipe(this);
     });
   }
   _write(a, b, c) {
@@ -303,24 +303,24 @@ class Ha extends va {
     return this.m;
   }
 }
-const Ia = async a => {
-  ({i:a} = new Ha({rs:a, u:R(!0)}));
+const Ja = async a => {
+  ({i:a} = new Ia({rs:a, u:R(!0)}));
   return await a;
 };
-async function Ja(a, b) {
-  return Ka(a, b);
+async function Ka(a, b) {
+  return La(a, b);
 }
-class La extends ua {
+class Ma extends va {
   constructor(a, b) {
     super(b);
-    this.i = (Array.isArray(a) ? a : [a]).filter(wa);
+    this.i = (Array.isArray(a) ? a : [a]).filter(xa);
     this.g = !1;
     this.j = b;
   }
   async replace(a, b) {
-    const c = new La(this.i, this.j);
+    const c = new Ma(this.i, this.j);
     b && Object.assign(c, b);
-    a = await Ja(c, a);
+    a = await Ka(c, a);
     c.g && this.brake();
     b && Object.keys(b).forEach(d => {
       b[d] = c[d];
@@ -351,7 +351,7 @@ class La extends ua {
             h instanceof Promise && e.push(h);
             return h;
           } catch (h) {
-            xa(f, h);
+            ya(f, h);
           }
         });
         if (e.length) {
@@ -359,7 +359,7 @@ class La extends ua {
             const k = await Promise.all(e);
             b = b.replace(c, () => k.shift());
           } catch (k) {
-            xa(f, k);
+            ya(f, k);
           }
         } else {
           b = g;
@@ -378,32 +378,32 @@ class La extends ua {
     }
   }
 }
-async function Ka(a, b) {
-  b instanceof ta ? b.pipe(a) : a.end(b);
-  return await Ia(a);
+async function La(a, b) {
+  b instanceof ua ? b.pipe(a) : a.end(b);
+  return await Ja(a);
 }
-;const Ma = a => {
+;const Na = a => {
   a = `(${a.join("|")})`;
   return new RegExp(`(\\n)?( *)(<${a}${"(?:\\s+(?!\\/>)[^>]*?)?"}(?:\\s*?/>|>[\\s\\S]*?<\\/\\4>))`, "gm");
 };
-var Na = tty;
-const {format:Oa, inspect:Pa} = util;
+var Oa = tty;
+const {format:Pa, inspect:Qa} = util;
 /*
 
  Copyright (c) 2016 Zeit, Inc.
  https://npmjs.org/ms
 */
-function Qa(a) {
+function Ra(a) {
   var b = {}, c = typeof a;
   if ("string" == c && 0 < a.length) {
-    return Ra(a);
+    return Sa(a);
   }
   if ("number" == c && isFinite(a)) {
     return b.C ? (b = Math.abs(a), a = 864E5 <= b ? S(a, b, 864E5, "day") : 36E5 <= b ? S(a, b, 36E5, "hour") : 6E4 <= b ? S(a, b, 6E4, "minute") : 1000 <= b ? S(a, b, 1000, "second") : a + " ms") : (b = Math.abs(a), a = 864E5 <= b ? Math.round(a / 864E5) + "d" : 36E5 <= b ? Math.round(a / 36E5) + "h" : 6E4 <= b ? Math.round(a / 6E4) + "m" : 1000 <= b ? Math.round(a / 1000) + "s" : a + "ms"), a;
   }
   throw Error("val is not a non-empty string or a valid number. val=" + JSON.stringify(a));
 }
-function Ra(a) {
+function Sa(a) {
   a = String(a);
   if (!(100 < a.length) && (a = /^(-?(?:\d+)?\.?\d+) *(milliseconds?|msecs?|ms|seconds?|secs?|s|minutes?|mins?|m|hours?|hrs?|h|days?|d|weeks?|w|years?|yrs?|y)?$/i.exec(a))) {
     var b = parseFloat(a[1]);
@@ -458,7 +458,7 @@ function S(a, b, c, d) {
  Copyright(c) 2015 Jed Watson
  MIT Licensed
 */
-const Sa = /\B(?=(\d{3})+(?!\d))/g, Ta = /(?:\.0*|(\.[^0]+)0+)$/, T = {b:1, kb:1024, mb:1048576, gb:1073741824, tb:Math.pow(1024, 4), pb:Math.pow(1024, 5)};
+const Ta = /\B(?=(\d{3})+(?!\d))/g, Ua = /(?:\.0*|(\.[^0]+)0+)$/, T = {b:1, kb:1024, mb:1048576, gb:1073741824, tb:Math.pow(1024, 4), pb:Math.pow(1024, 5)};
 function U(a, b) {
   if (!Number.isFinite(a)) {
     return null;
@@ -466,8 +466,8 @@ function U(a, b) {
   const c = Math.abs(a), d = b && b.D || "", e = b && b.G || "", f = b && void 0 !== b.s ? b.s : 2, g = !(!b || !b.B);
   (b = b && b.F || "") && T[b.toLowerCase()] || (b = c >= T.pb ? "PB" : c >= T.tb ? "TB" : c >= T.gb ? "GB" : c >= T.mb ? "MB" : c >= T.kb ? "KB" : "B");
   a = (a / T[b.toLowerCase()]).toFixed(f);
-  g || (a = a.replace(Ta, "$1"));
-  d && (a = a.replace(Sa, d));
+  g || (a = a.replace(Ua, "$1"));
+  d && (a = a.replace(Ta, d));
   return a + e + b;
 }
 ;/*
@@ -475,11 +475,11 @@ function U(a, b) {
  BSD License
  Copyright (c) 2009-2015, Kevin Decker <kpdecker@gmail.com>
 */
-const Ua = {black:30, red:31, green:32, yellow:33, blue:34, magenta:35, cyan:36, white:37, grey:90};
+const Va = {black:30, red:31, green:32, yellow:33, blue:34, magenta:35, cyan:36, white:37, grey:90};
 function V(a, b) {
-  return (b = Ua[b]) ? `\x1b[${b}m${a}\x1b[0m` : a;
+  return (b = Va[b]) ? `\x1b[${b}m${a}\x1b[0m` : a;
 }
-;var Va = {f:U, ["fy"](a) {
+;var Wa = {f:U, ["fy"](a) {
   return V(U(a) || "", "yellow");
 }, ["fr"](a) {
   return V(U(a) || "", "red");
@@ -501,13 +501,13 @@ const W = Object.keys(process.env).filter(a => /^debug_/i.test(a)).reduce((a, b)
 }, {}), X = {init:function(a) {
   a.inspectOpts = {...W};
 }, log:function(...a) {
-  return process.stderr.write(Oa(...a) + "\n");
+  return process.stderr.write(Pa(...a) + "\n");
 }, formatArgs:function(a) {
   const {namespace:b, useColors:c, color:d, diff:e} = this;
   if (c) {
     const f = "\u001b[3" + (8 > d ? d : "8;5;" + d), g = `  ${f};1m${b} \u001B[0m`;
     a[0] = g + a[0].split("\n").join("\n" + g);
-    a.push(f + "m+" + Qa(e) + "\u001b[0m");
+    a.push(f + "m+" + Ra(e) + "\u001b[0m");
   } else {
     a[0] = (W.hideDate ? "" : (new Date).toISOString() + " ") + b + " " + a[0];
   }
@@ -516,20 +516,20 @@ const W = Object.keys(process.env).filter(a => /^debug_/i.test(a)).reduce((a, b)
 }, load:function() {
   return process.env.DEBUG;
 }, useColors:function() {
-  return "colors" in W ? !!W.colors : Na.isatty(process.stderr.fd);
+  return "colors" in W ? !!W.colors : Oa.isatty(process.stderr.fd);
 }, colors:[6, 2, 3, 4, 5, 1], inspectOpts:W, formatters:{o:function(a) {
-  return Pa(a, {...this.inspectOpts, colors:this.useColors}).replace(/\s*\n\s*/g, " ");
+  return Qa(a, {...this.inspectOpts, colors:this.useColors}).replace(/\s*\n\s*/g, " ");
 }, O:function(a) {
-  return Pa(a, {...this.inspectOpts, colors:this.useColors});
-}, ...Va}};
-function Wa(a) {
+  return Qa(a, {...this.inspectOpts, colors:this.useColors});
+}, ...Wa}};
+function Xa(a) {
   function b(...g) {
     if (b.enabled) {
       var k = Number(new Date);
       b.diff = k - (f || k);
       b.prev = f;
       f = b.curr = k;
-      g[0] = Xa(g[0]);
+      g[0] = Ya(g[0]);
       "string" != typeof g[0] && g.unshift("%O");
       var p = 0;
       g[0] = g[0].replace(/%([a-zA-Z%]+)/g, (h, m) => {
@@ -550,20 +550,20 @@ function Wa(a) {
   let f;
   return b;
 }
-function Ya(a) {
-  const b = Wa(a);
+function Za(a) {
+  const b = Xa(a);
   "function" == typeof a.init && a.init(b);
   a.g.push(b);
   return b;
 }
-function Za(a, b) {
+function $a(a, b) {
   let c = 0;
   for (let d = 0; d < b.length; d++) {
     c = (c << 5) - c + b.charCodeAt(d), c |= 0;
   }
   return a.colors[Math.abs(c) % a.colors.length];
 }
-function $a(a) {
+function ab(a) {
   var b = X.load();
   a.save(b);
   a.i = [];
@@ -577,7 +577,7 @@ function $a(a) {
     b = a.g[c], b.enabled = a.enabled(b.namespace);
   }
 }
-class ab {
+class bb {
   constructor(a) {
     this.colors = a.colors;
     this.formatArgs = a.formatArgs;
@@ -614,18 +614,18 @@ class ab {
     return !1;
   }
 }
-function Xa(a) {
+function Ya(a) {
   return a instanceof Error ? a.stack || a.message : a;
 }
-;const {createWriteStream:bb} = fs;
-async function cb(a) {
-  var b = `export ${db.toString()}
+;const {createWriteStream:cb} = fs;
+async function db(a) {
+  var b = `export ${eb.toString()}
 
-export ${eb.toString()}`;
+export ${fb.toString()}`;
   if (!a) {
     throw Error("No path is given.");
   }
-  const c = R(!0), d = bb(a);
+  const c = R(!0), d = cb(a);
   await new Promise((e, f) => {
     d.on("error", g => {
       g = c(g);
@@ -633,23 +633,23 @@ export ${eb.toString()}`;
     }).on("close", e).end(b);
   });
 }
-;const {join:fb} = path;
-const gb = a => {
+;const {join:gb} = path;
+const hb = a => {
   const b = Object.keys(a);
   return `{
     ${b.map(c => `${/-/.test(c) ? `'${c}'` : c}: '${a[c]}'`).join(",\n")}${b.length ? "," : ""}
   }`;
-}, hb = a => {
+}, ib = a => {
   const b = [];
   b.push(`key: '${a.key}'`);
   b.push(`id: '${a.id}'`);
-  Object.keys(a.props).length && b.push(`props: ${gb(a.props)}`);
+  Object.keys(a.props).length && b.push(`props: ${hb(a.props)}`);
   a.children.filter(Boolean).length && b.push(`children: ${JSON.stringify(a.children)}`);
   return `{
 ${b.map(c => `  ${c}`).join(",\n") + ","}
 }`;
-}, ib = a => `/** @type {!Array<!preact.PreactProps>} */
-const meta = [${a.map(b => hb(b)).join(",\n")}]`, jb = (a = !0) => a ? `const io = makeIo(${"boolean" != typeof a ? JSON.stringify(a).replace(/([^\\])"([^"]+?)":/g, (b, c, d) => `${"," == c ? ", " : c}${d}: `).replace(/^{/, "{ ").replace(/}$/, " }") : ""})` : "", Y = a => a.replace(/(?:^|-)(.)/g, (b, c) => c.toUpperCase()), kb = a => `const __components = {\n  ${a.map(({key:b}) => `'${b}': ${Y(b)}`).filter((b, c, d) => d.indexOf(b) == c).join(",\n  ")},\n}`, lb = (a, b) => {
+}, jb = a => `/** @type {!Array<!preact.PreactProps>} */
+const meta = [${a.map(b => ib(b)).join(",\n")}]`, kb = (a = !0) => a ? `const io = makeIo(${"boolean" != typeof a ? JSON.stringify(a).replace(/([^\\])"([^"]+?)":/g, (b, c, d) => `${"," == c ? ", " : c}${d}: `).replace(/^{/, "{ ").replace(/}$/, " }") : ""})` : "", Y = a => a.replace(/(?:^|-)(.)/g, (b, c) => c.toUpperCase()), lb = a => `const __components = {\n  ${a.map(({key:b}) => `'${b}': ${Y(b)}`).filter((b, c, d) => d.indexOf(b) == c).join(",\n  ")},\n}`, mb = (a, b) => {
   const c = {};
   a.reduce((d, {key:e}) => {
     if (d.includes(e)) {
@@ -665,21 +665,22 @@ const meta = [${a.map(b => hb(b)).join(",\n")}]`, jb = (a = !0) => a ? `const io
   });
   return Object.entries(c).map(([d, e]) => Z(e, d));
 }, Z = (a, b, c = !0) => {
-  const [d, ...e] = a;
+  let [d, ...e] = a;
+  e = e.filter(Boolean);
   a = "import ";
   d && (a += c ? Y(d) : d);
   e.length && (a = a + (d ? ", " : "") + `{ ${(c ? e.map(Y) : e).join(", ")} }`);
   return a + ` from '${b}'`;
-}, db = require("./init"), eb = require("./make-io");
-var mb;
-mb = function() {
-  const a = new ab(X);
+}, eb = require("./init"), fb = require("./make-io");
+var nb;
+nb = function() {
+  const a = new bb(X);
   return function(b) {
-    const c = Ya(a);
+    const c = Za(a);
     c.namespace = b;
     c.useColors = X.useColors();
     c.enabled = a.enabled(b);
-    c.color = Za(a, b);
+    c.color = $a(a, b);
     c.destroy = function() {
       a.destroy(this);
     };
@@ -688,11 +689,11 @@ mb = function() {
       d.log = this.log;
       return d;
     };
-    $a(a);
+    ab(a);
     return c;
   };
 }()("competent");
-const nb = (a, b) => new RegExp(a.source.replace(new RegExp(`([|(])${b}([|)])`), (c, d, e) => "|" == d && "|" == e ? "|" : ")" == e ? e : "(" == d ? d : ""), a.flags);
+const ob = (a, b) => new RegExp(a.source.replace(new RegExp(`([|(])${b}([|)])`), (c, d, e) => "|" == d && "|" == e ? "|" : ")" == e ? e : "(" == d ? d : ""), a.flags);
 /*
 
  Competent: Render tags HTML as JSX components.
@@ -713,65 +714,72 @@ const nb = (a, b) => new RegExp(a.source.replace(new RegExp(`([|(])${b}([|)])`),
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 module.exports = {_competent:(a, b = {}) => {
-  async function c(t, y, Q, B, l, K, q) {
-    mb("render %s", l);
+  async function c(t, z, Q, C, l, L, q) {
+    nb("render %s", l);
+    const w = Error("Skip render");
     try {
-      const v = a[l], D = q.slice(0, K), r = q.slice(K + t.length);
-      if (/\x3c!--\s*$/.test(D) && /^\s*--\x3e/.test(r)) {
+      const x = a[l], r = q.slice(0, L), E = q.slice(L + t.length);
+      if (/\x3c!--\s*$/.test(r) && /^\s*--\x3e/.test(E)) {
         return t;
       }
-      const [{content:E = "", props:w}] = ma(l, B);
-      B = [E];
-      let z = !1, x = !0, F = !1, L, na, oa, pa, qa;
-      const ra = e.call(this, {...w, children:B}, {export(n = !0) {
-        z = n;
-      }, setPretty(n, I) {
-        L = n;
-        I && (na = I);
-      }, renderAgain(n = !0, I = !1) {
-        x = n;
-        F = I;
+      const [{content:A = "", props:y}] = ma(l, C);
+      C = [A];
+      let v = !1, I = !0, J = !1, na, oa, pa, qa, ra;
+      const sa = e.call(this, {...y, children:C}, {export(n = !0) {
+        v = n;
+      }, setPretty(n, H) {
+        na = n;
+        H && (oa = H);
+      }, renderAgain(n = !0, H = !1) {
+        I = n;
+        J = H;
       }, setChildContext(n) {
-        pa = n;
-      }, removeLine(n = !0) {
         qa = n;
+      }, removeLine(n = !0) {
+        ra = n;
+      }, skipRender() {
+        throw w;
       }}, l);
-      let A;
+      let B;
       try {
-        const n = v.call(this, ra);
-        A = n instanceof Promise ? await n : n;
+        const n = x.call(this, sa);
+        B = n instanceof Promise ? await n : n;
       } catch (n) {
         if (!n.message.startsWith("Class constructor")) {
           throw n;
         }
-        A = (new v).render(ra);
+        B = (new x).render(sa);
       }
-      z && !A.attributes.id && (oa = d.call(this), A.attributes.id = oa);
-      const sa = {pretty:L, lineLength:na};
-      let C;
-      "string" == typeof A ? C = A : Array.isArray(A) ? C = A.map(n => "string" == typeof n ? n : N(n, sa)).join("\n") : C = N(A, sa);
-      if (!C && qa) {
-        return g && g.call(this, l, w), "";
+      v && !B.attributes.id && (pa = d.call(this), B.attributes.id = pa);
+      const ta = {pretty:na, lineLength:oa};
+      let D;
+      "string" == typeof B ? D = B : Array.isArray(B) ? D = B.map(n => "string" == typeof n ? n : N(n, ta)).join("\n") : D = N(B, ta);
+      if (!D && ra) {
+        return g && g.call(this, l, y), "";
       }
-      C = (y || "") + C.replace(/^/gm, Q);
-      if (x) {
+      D = (z || "") + D.replace(/^/gm, Q);
+      if (I) {
         let n;
-        m ? n = m.call(this, l, F) : F ? n = {re:nb(u, l), replacement:c} : n = {re:u, replacement:c};
-        const I = new La(n);
+        m ? n = m.call(this, l, J) : J ? n = {re:ob(u, l), replacement:c} : n = {re:u, replacement:c};
+        const H = new Ma(n);
         if (h) {
-          const ob = h.call(this, pa);
-          Object.assign(I, ob);
+          const pb = h.call(this, qa);
+          Object.assign(H, pb);
         }
-        C = await Ja(I, C);
+        D = await Ka(H, D);
       }
-      z && f.call(this, l, A.attributes.id, w, B);
-      g && g.call(this, l, w);
-      return C;
-    } catch (v) {
-      return k && k.call(this, l, v, K, q), p ? "" : t;
+      v && f.call(this, l, B.attributes.id, y, C);
+      g && g.call(this, l, y);
+      return D;
+    } catch (x) {
+      if (x === w) {
+        return t;
+      }
+      k && k.call(this, l, x, L, q);
+      return p ? "" : t;
     }
   }
-  const {getId:d, getProps:e = (t, y) => ({...t, ...y}), markExported:f, onSuccess:g, onFail:k, removeOnError:p = !1, getContext:h, getReplacements:m} = b, u = Ma(Object.keys(a));
+  const {getId:d, getProps:e = (t, z) => ({...t, ...z}), markExported:f, onSuccess:g, onFail:k, removeOnError:p = !1, getContext:h, getReplacements:m} = b, u = Na(Object.keys(a));
   return {re:u, replacement:c};
 }, _makeComponentsScript:function(a, b) {
   if ("object" != typeof b) {
@@ -781,15 +789,15 @@ module.exports = {_competent:(a, b = {}) => {
   if (!c) {
     throw Error("The map of where to import components from is required.");
   }
-  var k = [Z([null, "render", ...f ? ["h"] : []], "preact", !1), ...d ? [Z([null, ...e ? ["makeIo"] : [], "init"], "./__competent-lib", !1)] : [], ...lb(a, c)].join("\n");
+  var k = [Z([null, "render", ...f ? ["h"] : []], "preact", !1), ...d ? [Z([null, ...e ? ["makeIo"] : [], "init"], "./__competent-lib", !1)] : [], ...mb(a, c)].join("\n");
   b = Object.keys(g).map(m => `props.${m} = ${g[m]}`).join("\n");
   const p = e ? "el.render = () => {\n    render(h(Comp, props, children), parent, el)\n  }\n  el.render.meta = { key, id }\n  io.observe(el)" : "render(h(Comp, props, children), parent, el)";
   k += "\n\n";
-  const h = kb(a);
+  const h = lb(a);
   k += h + "\n\n";
-  d || (k += db.toString() + "\n\n", e && (k += eb.toString() + "\n\n"));
-  e && (k += jb(e) + "\n\n");
-  k += ib(a);
+  d || (k += eb.toString() + "\n\n", e && (k += fb.toString() + "\n\n"));
+  e && (k += kb(e) + "\n\n");
+  k += jb(a);
   return k + `
 meta.forEach(({ key, id, props = {}, children }) => {
   const { parent, el } = init(id, key)
@@ -799,7 +807,7 @@ ${b ? `  ${b}` : ""}
 })
 `;
 }, _writeAssets:async a => {
-  await cb(fb(a, "./__competent-lib.js"));
+  await db(gb(a, "./__competent-lib.js"));
 }};